pipeline-ingestao/requirements.txt
python-dotenv
azure-core
azure-search-documents
openai
requests
beautifulsoup4